// src/types/api.ts
export interface components {
    schemas: {
// ...
        Model: {
            /** Id */
            id: string;
            /**
             * Created
             * @default 0
             */
            created: number;
            /**
             * Object
             * @default model
             * @constant
             */
            object: "model";
            /** Owned By */
            owned_by: string;
            /** Language */
            language?: string[] | null;
            /**
             * Task
             * @enum {string}
             */
            task: "automatic-speech-recognition" | "text-to-speech" | "speaker-embedding" | "voice-activity-detection";
        } & {
            [key: string]: unknown;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}